backend/app/requirements.txt
diffusers[torch]
fastapi
uvicorn
opencv-python
pydantic
numpy
onnx
onnxruntime-gpu
scikit-image
pillow